NF: move PreferenceUtils to the preferences package
